requirements.txt
requests
fastapi
uvicorn
pydantic-settings
python-jose